% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_api_docs.R
\docType{package}
\name{mojor-package}
\alias{mojor}
\alias{mojor-package}
\title{MöjoR package namespace directives.}
\description{
Core transpiler, JIT, and CPU runtime for MöjoR ("Numba for R").
}
\keyword{internal}
