% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_api_docs.R
\name{mojor_jit_signatures}
\alias{mojor_jit_signatures}
\title{List JIT Dispatcher Signatures}
\usage{
mojor_jit_signatures(
jit_fn,
kind = c("compiled", "declared", "all"),
format = c("key", "string", "typed_list")
)
}
\arguments{
\item{jit_fn}{Function returned by \code{\link[=mojor_jit]{mojor_jit()}}, \code{\link[=mojor_njit]{mojor_njit()}}, or \code{\link[=mojor_vectorize]{mojor_vectorize()}}.}

\item{kind}{Signature set to return (\code{"compiled"}, \code{"declared"}, or \code{"all"}).}

\item{format}{Output format (\code{"key"}, \code{"string"}, or \code{"typed_list"}).}
}
\value{
Signature set in requested format.
}
\description{
Return declared and/or compiled signatures from a JIT dispatcher in deterministic order.
}
